TW-828: Collectibles grid layout. Refactor
